requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
python-dotenv==1.0.0
pyyaml==6.0.2
pytest==8.3.0
httpx==0.27.0
